git-svn-id: http://ellcc.org/svn/ellcc/trunk@3029 82857169-6e38-49d7-aaa8-07fb558f12fc

+++ llvm/tools/lto/LTOModule.cpp
@@ -599,7 +599,7 @@ namespace {
         markGlobal(*Symbol);
     }
     virtual void EmitZerofill(const MCSection *Section, MCSymbol *Symbol,
-                              uint64_t Size , unsigned ByteAlignment) {
+                              unsigned Size , unsigned ByteAlignment) {
       markDefined(*Symbol);
     }
     virtual void EmitCommonSymbol(MCSymbol *Symbol, uint64_t Size,
